pyproject.toml
[project]
name = "extractive_summary"
version = "0.0.0"
description = "Validates that a string is a valid extractive summary of a given document."
authors = [
    {name = "Guardrails AI", email = "contact@guardrailsai.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.8"
dependencies = [
    "guardrails-ai>=0.3.2",
    "nltk",
    "pytest",
    "thefuzz"
]